pulling and pushing

+++ Bitmap-Engine/BitmapEngine/src/BBCCompressor.c
@@ -64,7 +64,7 @@ void bbcCompress(struct blockSeg *param){
 
     //0-fill byte or 1-fill byte (11111111 or 00000000)
     printf("param->header before first if %x\n", param->header);
-    
+
     //When a type 2 or type 4 is reached no new bytes can be added after the odd bit is placed so no matter what the new byte is a new run wll always be needed
     if(param->run_type == TYPE_2 || param->run_type == TYPE_4)
     {
@@ -162,7 +162,7 @@ void bbcCompress(struct blockSeg *param){
     //if the header is 0 and we're not on the last word of the block, start a new run.
     printf("param->header before second if :::: %x\n", param->header);
     //if the header is 0 and we're not on the last word of the block, start a new run.
-    if(param->header == 0 ){ //if any part of the above code caused a new run to be need this will start that new run //&& i != param->size-1 
+    if(param->header == 0 ){ //if any part of the above code caused a new run to be need this will start that new run //&& i != param->size-1
       printf("param->header was 0?\n");
       startNewRun(param);
     }
@@ -173,7 +173,7 @@ void bbcCompress(struct blockSeg *param){
   printf("param->header is %x\n", param->header);
   printf("param->size is %x\n", param->size);
   printf("writing final run, curr_size %x\n", param->curr_size);
-  printf("sizeof byte is %x\n", sizeof(byte));
+  printf("sizeof byte is %x\n",(unsigned int) sizeof(byte));
    printf("***************WRITING OUT**************** (file end) \n");
   fwrite(param->curr_run, sizeof(byte), param->curr_size+1, param->colFile);
   fclose(param->colFile);
